* commit 'ab3bb51f53ed0fb97dc4764c23d8abe088e959f8':
  AMDP-337 Fix page header
  AMDP-336 Fix border-top on the main menu of desktop version
  AMDP-325 Fix Page Header spacement
  AMDP-324 Fix page header of user profile page
  Fixing cleanup script
  Sudo for mkdir
  Creating mount point first
  Adding sudo
  Adding a script for deployment of secondary web hosts

+++ Deploy.sh
@@ -89,4 +89,6 @@ echo "Clear caches"
 sudo ${CWD}/drush --root=${APIGEE_DRUPAL_WEB_DOCROOT} cr
 
 #Delete old versions
-sudo rm -rf $APIGEE_DRUPAL_SOURCE_ROOT_RELEASE_OLD
+if [ "$APIGEE_DRUPAL_SOURCE_ROOT_RELEASE_OLD" != "$APIGEE_DRUPAL_SOURCE_ROOT_RELEASE" ]; then
+    sudo rm -rf $APIGEE_DRUPAL_SOURCE_ROOT_RELEASE_OLD
+fi
